hw/arm/boot: Assign dummy base address to memory node if needed
We possibly populate empty nodes where memory isn't included and might
be hot added at late time. The device-tree memory nodes can't be created
due to conflicting names, the 'unit-address' part precisely.

For example, the VM fails to start as the following error message indicates:

  /home/gavin/sandbox/qemu.main/build/qemu-system-aarch64          \
  -accel kvm -machine virt,gic-version=host                        \
  -cpu host -smp 4,sockets=2,cores=2,threads=1 -m 1024M,maxmem=64G \
  -object memory-backend-ram,id=mem0,size=512M                     \
  -object memory-backend-ram,id=mem1,size=512M                     \
  -numa node,nodeid=0,cpus=0-1,memdev=mem0                         \
  -numa node,nodeid=1,cpus=2-3,memdev=mem1                         \
  -numa node,nodeid=2                                              \
  -numa node,nodeid=3                                              \
    :
  -device virtio-balloon-pci,id=balloon0,free-page-reporting=yes

  qemu-system-aarch64: FDT: Failed to create subnode /memory@80000000: \
                       FDT_ERR_EXISTS

This fixes the issue by feeding dummy base address to the memory node
if it's a empty one. The memory node is still invalid as the 'length'
in 'reg' property is zero. In this way, we're following the device-tree
specification to include 'base address' as the 'unit-address' in the
memory node name, but the conflicting names can be avoided.

Signed-off-by: Gavin Shan <gshan@redhat.com>
---
v2: Used dummy base address instead of NUMA ID as the unit-adress.