hw/arm/virt: Move virt_flash_create() to machvirt_init()
For confidential VMs under CCA we have to skip the flash device creation
because they rely on read-only mappings, which we don't support. In
virt_instance_init() maching->cgs is not yet initialized, so we can't
know whether to skip the flash device creation or not. Move it later in
machvirt_init(), where we can access the machine->cgs member.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>