target/arm/kvm,tcg: Handle SMCCC hypercall exits in VMM during PSCI_CPU_{ON,OFF}
To support vCPU hotplug-like feature, we must trap any `HVC`/`SMC`
`PSCI_CPU_{ON,OFF}` hypercalls from the host KVM to QEMU for policy checks. This
ensures the following when a vCPU is brought online:

1. The vCPU is actually plugged in (i.e., present).
2. The vCPU is not administratively disabled. (Policy Checks)

Implement the registration and handling of `HVC`/`SMC` hypercall exits within
the VMM, ensuring that proper policy checks and control flow are enforced during
the vCPU onlining and offlining processes.

Co-developed-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Signed-off-by: Salil Mehta <salil.mehta@huawei.com>