debug: PCI hotplug
This tracks PCI hotplug on x86 and arm64 platform where the different
methods are used. On x86, the PCI hotplug is driven by ACPI (GPE)
events while that's driven by native event like MSIx interrupt on
arm64 platform. The goal of the debugging and investigation is to
understand the gap between those two mechanisms and to support ACPI
based PCI hotplug on arm64.

Signed-off-by: Gavin Shan <gshan@redhat.com>