HACK: hw/virtio/virtio-mem: Support guest-memfd-based RAM
When plugging memory into a confidential guest, it starts as
guest-private memory and doesn't need mappings in VFIO or vhost devices.
When the guest decides to share this memory, the RAM discard listeners
are notified. When the memory is unplugged, the guest first unshares the
memory if necessary, again notifying the RAM discard listeners. So we
don't need to install our own RAM discard listener.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>