target/arm/kvm: Support multiple memory CPERs injection
In the combination of 64KiB host and 4KiB guest, a problematic host
page affects 16x guest pages. In this specific case, it's reasonable
to push 16 consecutive memory CPERs. Otherwise, QEMU can run into core
dump due to the current error can't be delivered as the previous error
isn't acknoledged. It's caused by the nature that the host page can be
accessed by the guest in parallel due to the mismatched host and guest
page sizes.

Imporve push_ghes_memory_errors() to push 16x consecutive memory CPERs
for this specific case.

Signed-off-by: Gavin Shan <gshan@redhat.com>